requirements.txt
streamlit~=1.36.0
pylatexenc~=2.10
openai~=1.35.3
python-dotenv
htbuilder
anthropic